Fix GetGenericMethod and TestGenerics.
Now that we can specialize both the type and method, we need to fix
GetGenericMethod so we can find methods within specialized types.

Also, because this means the "DefiningType" for a method might be a
concretized generic (instead of a generic definition), we need to store
an actual TypeInfo reference in the MemberInfo.