Fix test failures in IDA generator script.
(1) We were having test failures because concrete generic arguments
were getting reference indices of -1. This was traced back to multiple
instantiations of a single TypeReference, likely because a TypeRef
refers to future TypeRefs, which were getting filled in and then
overwritten later in the model constructor's TypeRef loop.
(2) The official definition of ContainsGenericParameters considers the
ElementType too, which is important because otherwise T[] would not be
considered generic (this was causing an NPE in the IDA generator script
when it attempted to generate structs for T[]).