Incorporate enum ordinal in element name.
This makes the enum member name correspond exactly with the type class
name, which makes it possible to implement (janky) automatic TypeInfo
retyping through regex on the C source. (Again, only necessary for v16 -
for v19 and up the typing work automatically because we can grab the
references via MetadataUsages).