Quick Export #136
Allows you to export to the same directory of package without selecting path manually
And it will create folder if you export DLLs, C++ and VS solution
This might not be benefit for everyone, but I really like it and saves me some few clicks. Anyone who used my Il2CppDumper GUI version might find this useful :)

PS: I'm not good at WPF designing. I wanted the Quick Export to be placed beside and not on top of Export buttton